+++ src/main/groovy/org/labkey/gradle/plugin/Module.groovy
@@ -32,7 +32,7 @@ class Module implements Plugin<Project> {
     {
         project.apply plugin: 'org.labkey.build.javaModule'
 
-        if (!AntBuild.isApplicable(project) && BuildUtils.shouldDoBuild(project, false))
+        if (!AntBuild.isApplicable(project) && FileModule.shouldDoBuild(project, false))
         {
             addDependencies(project)
         }
