+++ src/main/groovy/org/labkey/gradle/task/UndeployModules.groovy
@@ -23,7 +23,6 @@ import org.gradle.api.tasks.TaskAction
 import org.labkey.gradle.plugin.FileModule
 import org.labkey.gradle.plugin.JavaModule
 import org.labkey.gradle.plugin.Module
-import org.labkey.gradle.util.BuildUtils
 
 /**
  * Removes modules from the deploy and staging directories.  If a value for dbType is provided,
@@ -40,7 +39,7 @@ class UndeployModules extends DefaultTask
     {
         project.rootProject.allprojects.each { Project p ->
             if (isLabKeyModule(p) &&
-                    (dbType == null || !BuildUtils.shouldDoBuild(p, true) || !JavaModule.isDatabaseSupported(p, dbType)))
+                    (dbType == null || !FileModule.shouldDoBuild(p, true) || !JavaModule.isDatabaseSupported(p, dbType)))
             {
                 this.logger.info("Undeploying module ${p.path} for dbType ${dbType}")
                 FileModule.undeployModule(p)
